requirements.txt
livekit-plugins-noise-cancellation~=0.2
python-dotenv
livekit-agents
livekit-agents[google]~=1.2
mem0ai
duckduckgo-search
langchain_community
requests